refactor(body-punctuation): 各定数を使用箇所へ寄せてsharedを廃止
`shared.ts`には文字集合・閾値・終端正規表現をまとめて置いていましたが、
`commaChars`と`commaPrefixThreshold`は読点判定でしか使わず、
`periodChars`は句点判定でしか使わず、
`defaultTerminator`は行末判定でしか使わない、
というように実際はそれぞれ単一の箇所からしか参照されていませんでした。
共通定数のように見えてしまい関心の分離の観点で素直ではないため、
各定数を実際に使うファイルへ移します。

結果として`shared.ts`は`extractLines`の再エクスポートのみになり、
存在意義が薄れるため廃止し、
各ルールは`extract-markdown`から直接`extractLines`をimportします。
